refactor: improve home component organisation, declare home width in parent instead of children
