--- src/components/MainClaim.tsx
@@ -1,13 +1,0 @@
-export default function MainClaim() {
-  return (
-    <div className='text-main-mobile lg:text-main px-10 font-mono font-extrabold text-white lg:px-28'>
-      <span className='text-white/50'> UNLOCK YOUR </span> POTENTIAL
-      <br />
-      <span className='text-white/50'> WITH OUR </span> FULL STACK DEVELOPMENT
-      <br className='hidden sm:inline-flex' />
-      <span className='text-white/50'> AND </span>{' '}
-      <br className='inline-flex sm:hidden' />
-      IT CONSULTING
-    </div>
-  )
-}
