+++ src/components/home/WhatWeDo.tsx
